Удаление фильма и пользователя по id - основа

+++ src/main/java/ru/yandex/practicum/filmorate/storage/db/UserDbStorage.java
@@ -65,10 +65,9 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public String delete(int userId) {
-        String sqlQuery = "DELETE FROM users WHERE user_id = ?";
-        jdbcTemplate.update(sqlQuery, userId);
-        return sqlQuery;
+    public void delete(int userId) {
+        String sqlQuery = "DELETE FROM users WHERE user_id = " + userId;
+        jdbcTemplate.update(sqlQuery);
     }
 
     public User getById(Integer userId) {
@@ -94,6 +93,7 @@ public class UserDbStorage implements UserStorage {
     }
 
     public List<User> getFriends(int userId) {
+        getById(userId);
         List<User> friends = new ArrayList<>();
         String sqlQuery = "SELECT * FROM users "
                 + "WHERE users.user_id IN (SELECT friend_id from friends "
